Refactor: move session helpers from endpoints to utils
This removes an architectural inversion where api.utils depended on
api.endpoints. The session validation functions (is_session_too_old,
force_user_to_relog) are now in utils/session_helpers.py where they
belong, since they're utility functions rather than endpoint handlers.
